Add sh scripts for edge cases

+++ prepare_trip_dataset.py
@@ -48,22 +48,26 @@ if __name__ == '__main__':
         print("Please set --path")
         sys.exit()
     else:
-        for filename in os.listdir(args.path):
-            f = os.path.join(args.path, filename)
-            checkCSV(f)
-            print("load trip dataset...", filename)
-            df = pd.read_csv(f, skipinitialspace=True, usecols=col_list, low_memory=False)
+        p = 'data/trips/' + args.name + "/"
+        if not os.path.isfile(p + "trips.csv"):
+            for filename in os.listdir(args.path):
+                f = os.path.join(args.path, filename)
+                checkCSV(f)
+                print("load trip dataset...", filename)
+                df = pd.read_csv(f, skipinitialspace=True, usecols=col_list, low_memory=False)
 
-            print("drop unnecessary columns...", filename)
-            df.drop(['medallion', 'hack_license', 'dropoff_longitude', 'dropoff_latitude', 'dropoff_datetime', 'vendor_id', 'rate_code', 'store_and_fwd_flag', 'trip_time_in_secs', 'trip_distance'], axis=1, inplace=True)
+                print("drop unnecessary columns...", filename)
+                df.drop(['medallion', 'hack_license', 'dropoff_longitude', 'dropoff_latitude', 'dropoff_datetime', 'vendor_id', 'rate_code', 'store_and_fwd_flag', 'trip_time_in_secs', 'trip_distance'], axis=1, inplace=True)
 
-            print("get trips in topology bounds...")
-            df = df[df['pickup_longitude'].between(min_lon, max_lon)]
-            df = df[df['pickup_latitude'].between(min_lat, max_lat)]
-            df.sort_values("pickup_datetime", inplace=True)
-            p = 'data/trips/' + args.name + "/"
-            os.makedirs(p, exist_ok=True)
-            df.to_csv(p + filename)
-        df = pd.concat(map(pd.read_csv, glob.glob(os.path.join("data/trips/" + args.name + "/", "*.csv"))))
-        df.to_csv("data/trips/" + args.name + "/trips.csv")
-
+                print("get trips in topology bounds...")
+                df = df[df['pickup_longitude'].between(min_lon, max_lon)]
+                df = df[df['pickup_latitude'].between(min_lat, max_lat)]
+                df.sort_values("pickup_datetime", inplace=True)
+                os.makedirs(p, exist_ok=True)
+                df.to_csv(p + filename)
+            df = pd.concat(map(pd.read_csv, glob.glob(os.path.join(p, "*.csv"))))
+            df.to_csv(p + "trips.csv")
+        for filename in os.listdir(p):
+            if filename != "trips.csv":
+                f = os.path.join(p, filename)
+                os.remove(f)
